SVN revision 13221

+++ src/common/misc/pint-malloc.c
@@ -215,12 +215,12 @@ static int memdebugflag = 0;
 typedef struct extra_s
 {
     void     *mem;
-    size_t   size;
+    int64_t   size;
 #if PVFS_MALLOC_MAGIC
-    uint32_t magic;
+    uint64_t magic;
 #endif
 #if PVFS_MALLOC_CHECK_ALIGN
-    size_t   align;
+    int64_t   align;
 #endif
 } extra_t;
 
@@ -355,7 +355,7 @@ int PINT_check_malloc(void *ptr)
     {
         return 0;
     }
-    if (extra->magic == PVFS_MALLOC_MAGIC_NUM)
+    if (extra->magic == (uint64_t)PVFS_MALLOC_MAGIC_NUM)
     {
         return 1;
     }
@@ -396,7 +396,7 @@ void *PINT_malloc (size_t size __PMDBG)
 #endif
     extra->mem   = mem;
 #if PVFS_MALLOC_MAGIC
-    extra->magic = PVFS_MALLOC_MAGIC_NUM;
+    extra->magic = (uint64_t)PVFS_MALLOC_MAGIC_NUM;
 #endif
     extra->size  = sizeplus;
 #if PVFS_MALLOC_CHECK_ALIGN
@@ -458,7 +458,7 @@ int PINT_posix_memalign (void **mem, size_t alignment, size_t size __PMDBG)
 #endif
     extra->mem   = mem_orig;
 #if PVFS_MALLOC_MAGIC
-    extra->magic = PVFS_MALLOC_MAGIC_NUM;
+    extra->magic = (uint64_t)PVFS_MALLOC_MAGIC_NUM;
 #endif
     extra->size  = sizeplus;
 #if PVFS_MALLOC_CHECK_ALIGN
@@ -522,7 +522,7 @@ void *PINT_realloc (void *mem, size_t size __PMDBG)
 
     extra = (void *)((ptrint_t)mem - EXTRA_SIZE);
 #if PVFS_MALLOC_MAGIC
-    if (extra->magic != PVFS_MALLOC_MAGIC_NUM)
+    if (extra->magic != (uint64_t)PVFS_MALLOC_MAGIC_NUM)
     {
         gossip_err("PINT_realloc: realloc fails magic number test\n");
         gossip_err("mem = %p size = %d, emem = %p, esize = %d\n",
@@ -623,7 +623,7 @@ void PINT_free (void *mem __PMDBG)
     memdebug(dbfp, "\n");
 
 #if PVFS_MALLOC_MAGIC
-    if (extra->magic != PVFS_MALLOC_MAGIC_NUM)
+    if (extra->magic != (uint64_t)PVFS_MALLOC_MAGIC_NUM)
     {
 #if PVFS_MALLOC_DEBUG
         gossip_lerr("PINT_free: free fails magic number test (%s line %d)\n",
